% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highofficer.R
\docType{data}
\name{highofficer}
\alias{highofficer}
\title{High Ranking officer Property Data in S.Korea 2018-2019}
\format{Tibble with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}
\item{carrier}{Two letter abbreviation}

}}
\source{
https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236
}
\usage{
highofficer
}
\description{
Every ends of March, Korea Government offer data of High Ranking officer Property Data in S.Korea 2018-2019
}
\examples{
  highofficer
}
\keyword{datasets}
