% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taltal-package.R
\docType{package}
\name{taltal-package}
\alias{taltal}
\alias{taltal-package}
\title{taltal: High Ranking officer Property Data in S.Korea 2018-2019}
\description{
Privide High Ranking officer Property Data in S.Korea 2018-2019 from official gazette
}
\author{
\strong{Maintainer}: Yeowoon Bae \email{bae.yeowoon@joongang.co.kr}

}
\keyword{internal}
