% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highofficer.R
\docType{data}
\name{highofficer}
\alias{highofficer}
\title{High Ranking officer Property Data in S.Korea 2018-2019}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 226 rows and 16 columns.}
\source{
https://news.joins.com/DigitalSpecial/349
}
\usage{
highofficer
}
\description{
Every ends of March, Korea Government offer data of High Ranking officer Property Data in S.Korea 2018-2019
}
\examples{
  highofficer
}
\keyword{datasets}
